requirements.txt
aiohttp==3.9.1
spacy==3.7.2
streamlit==1.29.0
pyyaml==6.0.1
psycopg2==2.9.9